Fix wrong restore of scroll position when navigating back from detail
When going back to the list from a movie detail, the scroll position is not restored properly if the list was past the 1st page.
This seems to be because the list will have to page again starting from the 1st page so the scroll position to restore won't be there at first.
Setting `enablePlaceholders = true` will make the pager emit placeholder items for items in the list which have been fetched but have not yet been added to the pager.
This way the scroll position can be restored proeprly by the lazy column.